requirements/development.txt
-r base.txt
django-debug-toolbar==5.1.0
